Рефакторинг handlers.py
