+++ mimir/handlers/__init__.py
@@ -1,0 +1,25 @@
+from .init import handle_init
+from .task import handle_create_task, handle_list_tasks
+from .commit import handle_commit, handle_history, handle_show
+from .branch import handle_branch, handle_branch_list, handle_branch_create, handle_branch_delete
+from .context import handle_context, handle_switch
+from .project import handle_create_project, handle_list_projects
+from .status import handle_status
+
+__all__ = [
+    "handle_init",
+    "handle_create_task",
+    "handle_list_tasks",
+    "handle_commit",
+    "handle_history",
+    "handle_show",
+    "handle_branch",
+    "handle_branch_list",
+    "handle_branch_create",
+    "handle_branch_delete",
+    "handle_context",
+    "handle_switch",
+    "handle_create_project",
+    "handle_list_projects",
+    "handle_status",
+]
